chore: renamed cmd to core
