+++ core/api/generator.go
