+++ core/aug/base.go
