+++ core/fullstack/config.go
