+++ core/fullstack/generator.go
