+++ core/opts/package.go
