+++ core/ui/generator.go
